Day 7 (3656/3374)
I started out just representing all paths as strings (e.g. /a/e/i).
Re-reading the problem, I realized that directory and file sizes should
contribute recursively to each parent directory (duh, it's a file
system). So I ended up pivoting to a tree-like structure in the middle
of it all, when I should have done that from the beginning.

That combined with a ton of buggy string handling and tree walks made
for a pretty horrible performance...

Day 1 of https://adventofrealizingicantread.com