2015/4: Compute eight hashes in parallel with AVX2
This makes it 3-4 times faster than before on my 3900X; running it on my
input now takes around 25 ms.

There might still be a little bit to gain by manually vectorizing the
input reshuffling in md5_block_avx2(). GCC does auto-vectorize it, but
the assembly looks a bit messy. I'm done with this for now though...