2015/4: Implement a faster version of to_chars()
This implementation is more or less the same idea as used in libstdc++,
writing two digits at a time.

One modification is that instead of determining the number of digits by
repeated division, we indirectly derive it from the base-2 logarithm of
the input, by using the leading zero count as an table index. This takes
constant time with regards to the order of magnitude of the input.

It does have the trade-off of requiring an additional ~500 bytes of
constant data, and potentially the additional cache misses caused by
this. For this application, however, it seems to be slightly faster; I
assume that everything fits in cache in this case.

Before:

  Year    Day    Mean (μs)    σ (μs)    10% (μs)    90% (μs)
------  -----  -----------  --------  ----------  ----------
  2015      4      17763.8   5234.67     13838.7     23162.4

After:

  Year    Day    Mean (μs)    σ (μs)    10% (μs)    90% (μs)
------  -----  -----------  --------  ----------  ----------
  2015      4      16811.3   5011.39     12873.6     21635.9