2023/1: Speed up
Instead of comparing individual characters in each position, which leads
to O(m*n) behaviour, hold the next eight characters in the string in a
64-bit integer as a rolling window. Masking and comparing this integer
to an expected string packed in the same way is O(1) which brings the
entire loop back down to O(n) complexity.

(Also, the logic is now completely inlined which avoids the expensive
call out to string_view::compare() and memcmp() for every character.)

I really can't be bothered to implement Aho-Corasick for a few μs gain
over this (at best).

  Year    Day    Mean (μs)    σ (μs)    10% (μs)    90% (μs)
------  -----  -----------  --------  ----------  ----------
  2023      1      94.2895   2.14888      93.557       94.91