common: Optimize getline() further
It now reads directly into the buffer of the string, reducing the number
of copies of the data by one.

The drawback is that we now have to find the newline that fgets() placed
in the buffer, if there was one, to be able to resize the string to
discard the newline and anything after it. Look at 32 bytes at a time
using AVX.

All this because fgets() doesn't give us the number of bytes that it
wrote to the buffer. (fread would be nicer here - or better yet,
handling all I/O buffering ourselves - but I'm too lazy to go back and
update all the problems that I've written so far.)