dense_map: Find occupied buckets with SSE
Vectorize this, looking at 16 buckets at a time to find the next empty
one. Looking at 32 buckets at once with AVX seems to be slightly slower,
likely because most hash tables aren't large enough to benefit from that
to begin with.

Bucket states now signify whether iterators should stop at that bucket
with the top bit (0x80) rather than the bottom bit. This allows for very
quick checking since we can load 16 states and then immediately use
movemask to get the offset of the next occupied/sentinel bucket.

Add 16 sentinel values instead of a single one to the end of the bucket
state array to avoid having to fall back to a scalar loop for the
elements near the end of the table.