2023/17: Optimize
Don't store the number of forward steps in each graph vertex. Instead,
let the neighbor function generate multiple steps in the same direction,
respecting the minimum and maximum number of such steps, assuming that
we will turn at the end of that line.

This helps to greatly reduce the size of the state space since we are
now dealing with many duplicate (position, dir) pairs that can easily be
culled, rather than (position, dir, n_steps) tuples which are often
unique and hence have to be visited.

  Year    Day    Mean (μs)    σ (μs)    10% (μs)    90% (μs)
------  -----  -----------  --------  ----------  ----------
  2023     17      42770.6    721.95     42395.1     43138.6