dense_map: Fix insertion with an rvalue key when rehashing
Rehashing currently requires us to find the correct bucket for the key
again, since it might have moved. However, constructing the entry with
an rvalue key before the rehash check will leave it in a moved-from
state, which we then pass to find_bucket_().

Fix this by rehashing before actually inserting the entry.