dense_map: Store 6 bit hash in bucket state
Since we have 6 bits to spare for each bucket state, place the lowest 6
bit of the hash there. When searching for a key, we can use this to cut
down the equality checks for non-matching elements to 1/64 of elements
in a probe chain on average, assuming the hashes are equidistributed.