dense_map: SIMD bucket probing
Vectorize the bucket probing to check 32 buckets at a time using AVX2.
For the vast majority of cases (assuming a good hash function), a single
iteration of this loop is enough to either find the correct bucket or
deduce that it does not exist with only a single equality check.