2019/intcode: Fix multiplication overflows
In the merged OP_ADD and OP_MUL cases, both the sum and product were
computed beforehand to encourage the compiler to emit instructions for
both (add/lea and imul) and then select the correct result with a cmov.
This is usually a small win as both can be computed cheaply in parallel,
and which of the two operation we are handling at any point in time is
usually unpredictable.

However, this can run afoul of undefined behaviorw: for some of my 2019
inputs, we end up executing an OP_ADD where sum is well-defined, but the
product overflows, even with 64-bit integers. The emitted code is fine,
since imul overflowing doesn't matter on x86-64 -- we just get garbage
that is ignored by cmov -- but in C++ we are invoking nasal demons.

Inline the sum and product into the ternary operator to avoid evaluating
the multiplication unless we are executing an OP_MUL. GCC 15.1.1 emits a
lea+imul followed by a cmov, so the buggy construction appears to have
been pointless to begin with.