common: Optimize find_numbers_impl()
If we are dealing with unsigned numbers, there's no point in even
looking for a negative sign. Even if we _do_ want to look for one, it's
faster to just pretend that we are looking for an unsigned number and
only go back to check if there was a minus sign at the very end.

Also replace std::from_chars() with a simpler loop which only performs
overflow checking in debug mode.